requirements.txt
# Telegram bot framework
python-telegram-bot[job-queue]==21.5

# HTTP client (async)
httpx==0.27.2

# HTML parsing
beautifulsoup4==4.12.3

# Environment variables
python-dotenv==1.0.1

# Timezone support (for scheduler)
pytz==2024.1

feedparser==6.0.11
